Add sample filters, show sample name first, add sample link, fix genome link, fix sample table count text
